// app-fundacion-prod01/src/app/components/model/user.model.ts
export class User { 
    role: string;
    tipodocumento: string;
    documento: string;
    email: string;
    password: string;
    histpassword: string;
    passwordcheck:string;
    estado: string;
    nombre: string;
    telefono:string;
    rol:string;
    maestro:string;
    permission:string;
    maestropermission : string;
    creadopor: string;
    creadoen: string;
    modificadopor: string;
    modificadoen: string;
    modificacioncausa:string;
  }