// app-fundacion-prod01/src/app/components/model/garantias.model.ts
export class Garantias {
    indicativo: string;
    confirmacion: string;
    idcolecciongarantias: string;
    idcoleccion: string;
    idcoleccionpadre:string;
    idgarantiaregistro: string;
    idgarantia: string;
    idcomision: string;
    idcomisionregistro: string;
    idlinea: string;
    intermediario: string;
    intermediariodes:string;
    codigolineacrediticio: string;
    tipoidentificacion: string;
    identificacion;
    codigociiu: string;
    apellidos: string;
    nombres: string;
    municipiocliente: string;
    direccionclienteuno: string;
    direccionclientedos: string;
    telefonouno: string;
    telefonodos: string;
    celular: string;
    correoelectronico: string;
    nombrecodeudor: string;
    identificacioncodeudor: string;
    direccioncodeudor: string;
    telefonocodeudor: string;
    celularcodeudor: string;
    credito: string;
    pagare: string;
    plazo: string;
    periodo: string;
    tasa: string;
    fechadesembolso: string;
    fechavencimiento: string;
    amortizacion: string;
    periodogracia: string;
    valormontodesembolsado: string;
    saldo: string;
    saldorestante: string;
    saldocobrado;
    cobertura: string;
    administracion: string;
    ivaadministracion: string;
    comisiontotal: string;
    comisiontotaltotal: string;
    estado: string;
    creadopor: string;
    creadoen: string;
    modificadopor: string;
    modificadoen: string;
    modificacioncausa: string;
    pagarefile: string;
    cartainstruccion: string;
    liquidacioncredito: string;
    cedulaciudadania: string;
    liquidaciongarantia: string;
    pagareurlFileUpload;
    cartainstruccionurlFileUpload;
    liquidacioncreditourlFileUpload;
    cedulaciudadaniaurlFileUpload;
    liquidaciongarantiaurlFileUpload;
    certificacionriesgo: string;
    anexoservicio: string;
    anexoserviciourlFileUpload;
    certificacionriesgourlFileUpload;
    validatemen;
    validatemenpag;
    registros;
}